Wikipedia::revisions - specify `rvslots`
Per https://www.mediawiki.org/wiki/API:Revisions `rvslots` being omitted
will cause data to be returned in a backwards-compatible format, however
is deprecated.

Previously when `content` was requested the response would be structured
similar to
```
{
  "query": {
    "pages": {
      "xxx": {
        "revisions": [
          {
            "*": "Page content"
          }
        ]
      }
    }
  }
}
```

With the slot specified, the content is re-structured similar to
```
{
  "query": {
    "pages": {
      "xxx": {
        "revisions": [
          {
            "slots": {
              "main": {
                "*": "Page content"
              }
            }
          }
        ]
      }
    }
  }
}
```

Thus also update the consumers to look at the new structure.