health check - skip authentication
When `Wikipedia\Http` is constructed there is a cookie jar created to
enable persistent sessions. The path of this cookie jar is randomised to
prevent different processes from conflicting.

As the health check script is run outside of the main bot process, it
gets a unique cookie jar setup.

The health check is run by k8s every 10 seconds, which results in a new
authentication to the API. We don't actually need authentication for the
API calls we execute, so skip the login stage to save resources.