processEdit - kill the child
Currently when the user is whitelisted, the function is returned,
however this happens after we have forked, resulting in many child
processes being created and the child processes processing multiple
edits.

Move the whitelist check before the fork, this logic is quite fast to
execute and we already do work pre-fork today.

This will both save resources (not needing a very short lived child) and
resolve the current issue