Re-enable spam feed
Currently we rely on the spam feed for huggle.

The best option is to refactor the data into a structured form, then
handle this in the message dispatching logic.

That will let us de-couple the spam feed from the 'data feed' and enable
more-reliable message formatting.

Once done we can flip this flag again.